% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normed.R
\name{normv}
\alias{normv}
\title{Calc. normalization of numeric vector using medium}
\usage{
normv(v)
}
\arguments{
\item{v}{a numeric vector}
}
\value{
The medium normalized vector of \code{v}
}
\description{
Calc. normalization of numeric vector using medium
}
\examples{
normv(c(36,7,111, 43, 84))

}
\author{
P. Terry
}
\seealso{
\code{median}
}

